Extract representation-dependent parts of FieldElement to a separate file
And compile it as a feature
